requirements.txt
langgraph>=0.2.0
google-genai>=0.3.0
tavily-python>=0.3.0
pydantic>=2.0.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0
langchain-core>=0.3.0
langchain-community>=0.3.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
# FastAPI and Authentication
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6